napisan controller za citanje podataka iz baze, nije testirano

+++ Models/UciniciClass.cs
@@ -1,9 +1,14 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
 namespace webSeminarski.Models{
 
     public class Ucenici{
         public int id {get;set;}
         public string ime {get;set;}
         public string prezime {get;set;}
-        public float prosek {get;set;}
+        public double prosek {get;set;}
     }
 }
